Move read_selection() to channels.read_neuromag_selection()
Fixes #8865
