+++ tutorials/simulation/plot_dics.py
@@ -187,7 +187,8 @@ assert len(epochs) == 2  # ensure that we got the two expected events
 
 # Plot some of the channels of the simulated data that are situated above one
 # of our simulated sources.
-picks = mne.pick_channels(epochs.ch_names, mne.read_selection('Left-frontal'))
+picks = mne.pick_channels(epochs.ch_names,
+                          mne.read_neuromag_selection('Left-frontal'))
 epochs.plot(picks=picks)
 
 ###############################################################################
