+++ tutorials/stats-sensor-space/plot_stats_cluster_1samp_test_time_frequency.py
@@ -54,7 +54,7 @@ epochs = mne.Epochs(raw, events, event_id, tmin, tmax, picks=picks,
                     baseline=(None, 0), preload=True,
                     reject=dict(grad=4000e-13, eog=150e-6))
 # just use right temporal sensors for speed
-epochs.pick_channels(mne.read_selection('Right-temporal'))
+epochs.pick_channels(mne.read_neuromag_selection('Right-temporal'))
 evoked = epochs.average()
 
 # Factor to down-sample the temporal dimension of the TFR computed by
